refactored network constants to own class
