+++ core/src/main/java/com/mankomania/game/core/network/NetworkConstants.java
@@ -1,0 +1,14 @@
+package com.mankomania.game.core.network;
+
+/*********************************
+ Created by Fabian Oraze on 16.04.20
+ *********************************/
+
+public class NetworkConstants {
+
+    public static final int TCP_PORT =  53211;
+    public static final String IP_HOST = "se2-demo.aau.at";
+    public static final int TIMEOUT = 5000;
+
+
+}
